src/requirements.txt
# requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.5.0
cryptography==41.0.3
python-dotenv==1.0.0